requirements.txt
pytelegrambotapi~=4.29.1
python-dotenv~=1.2.1
flask~=3.1.2
gspread~=6.2.1
google~=3.0.0